+++ src/components/LinkChooser/SearchResultSection.tsx
@@ -1,6 +1,8 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
 import { useListBoxSection } from "@react-aria/listbox";
 import React, { FC } from "react";
-import { SearchResultSectionProps } from ".";
+import { SearchResultSectionProps } from "./types";
 import { SearchResultOption } from "./SearchResultOption";
 
 export const SearchResultSection: FC<SearchResultSectionProps> = ({
